requirements.txt
python-telegram-bot==20.7
pytube==15.0.0
youtube-dl==2021.12.17
yt-dlp==2023.10.13
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.1
python-dotenv==1.0.0
Flask==2.3.3
gunicorn==21.2.0